media-source/MediaSource-removeSourceBuffer_removesourcebuffer.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>removesourcebuffer event</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var video = document.createElement('video');
var mimes = ['video/webm; codecs="vorbis,vp8"', 'video/mp4'];

//check the browser supports the MIME used in this test
function isTypeSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function mediaTest(mime) {
    async_test(function(t) {
        if(!isTypeSupported.bind(t)(mime)) {
            return;
        }
        var mediaSource = new MediaSource();
        var events = [];
        mediaSource.addEventListener('sourceopen', function(e) {
            events.push('sourceopen');
            var sourceBuffer = this.addSourceBuffer(mime);
            this.removeSourceBuffer(sourceBuffer);
        });
        mediaSource.activeSourceBuffers.addEventListener('removesourcebuffer', function(e) {
            events.push('removesourcebuffer by activeSourceBuffers');
        });
        mediaSource.sourceBuffers.addEventListener('removesourcebuffer', function(e) {
            events.push('removesourcebuffer by sourceBuffers');
            assert_equals(events.length, 3);
            t.step(function() {
                assert_array_equals(events, ['sourceopen',
                                             'removesourcebuffer by activeSourceBuffers',
                                             'removesourcebuffer by sourceBuffers']);
            });
            t.done();
        });
        video.src = window.URL.createObjectURL(mediaSource);
    }, 'MediaSource#removeSouceBuffer (' + mime + ') : ' +
       'If sourceBuffer is in activeSourceBuffers, then remove sourceBuffer from activeSourceBuffers ' +
       'and queue a task to fire a simple event named removesourcebuffer at the SourceBufferList ' +
       'returned by activeSourceBuffers. Remove sourceBuffer from sourceBuffers and queue a task to ' +
       'fire a simple event named removesourcebuffer at the SourceBufferList returned by sourceBuffers.');
}
mimes.forEach(function(mime) {
    mediaTest(mime);
});
</script>
</body>
</html>
